inlining field, formalizeable value to be more consistent with other fields, conclusions
